Remove tests from run configuration. (They don't work with Gradle in IntelliJ throwing strange errors). Renamed the tree view classes to match the library naming. Added documentation for tree view classes and new methods of NodeUtils.
ISelectionModel: added scanTree method, removed useless boolean from clearSelection method.

MFXTreeItem: Added css stylesheet for MFXTreeItem, it fixes a weird layout bug that occurres when the item's border is set to whatever color.
Added new constructor and moved defaultCellFactory method to the first constructor which doesn't specify the cell factory.
Changed layout strategy, it's more correct to use the VBox margin instead of setting the padding.
Removed the buildEventDispatchChain as after testing it seems unnecessary.
Added a new event type: ADD_REMOVE_ITEM_EVENT which is fired to adjust the layout accordingly when one or more items are added/removed.

MFXTreeItemSkin: added fade in and fade out animations for all children items when expand/collapse, thus switched to a ParallelTransition
Fixed the layout when items are added/removed and the item is either expanded or set to start expanded.
Switched to event filter on disclosure node because we want to intercept the event as soon as possible, so we can decide if the event should be consumed or executed.

MFXSimpleTreeCell: use convenience class MFXIconWrapper for the disclosure node.

NodeUtils added more convenience methods.

Signed-off-by: PAlex404 <alessandro.parisi406@gmail.com>