New control: MFXTableView
Added ScenicView to demo module for debugging purposes.

Corrected names for TreeCheckModel and TreeSelectionModel.
MFXStageDialog: it's now possible to block drag and set the manual position of the Stage.
RippleGenerator: fixed memory leak, the shape used for the background animation must also be removed at the end of the animation.

MFXComboBoxSkin: close the popup when the focus is lost.
Use computed sizes and unbind the properties in case maxPopupHeightProperty and maxPopupWidthProperty are set to -1.

MFXLabelSkin: minWidth should be correctly calculated now.

NodeUtils: added new method to calculate the minWidth of a Label given the Font and the text.
StringUtils: added new method to check if a CharSequence contains another CharSequence ignoring case.

mfx-scrollpane.css: fixed viewport background color.

Added IFilterable, MFXFilterDialog for filtering MFXTableView, note that the interface is optional, in case the Object doesn't implement it, toString will be used instead.
Added selection model for MFXTableView.
Added DragResizer for resizing nodes at runtime using the cursor.

Signed-off-by: PAlex404 <alessandro.parisi406@gmail.com>