Fix MFXComboBox sometimes triggering in 1 click instead of 2
MFXComboBox can be triggered(opened/closed) by only one click after shortly being opened/closed.

Steps to reproduce: open the ComboBox using a double click then then click once and the combo will close. This inconsistency can lead for unintended behavior in this situation as the user may want to close the ComboBox and thus will initiate a double click but the ComboBox will close and reopen instead of simply closing.

To mitigate it, i added a %2 check to only trigger based on a double click.