MFXComboBox - Fixed repeated click on arrow to break the behavior of opening/closing.
When clicking the arrow repeatedly, the functionality will be broken due to the event also triggering the double click logic.

Steps to reproduce: run demo and repeatedly click on the arrow, the combobox  will remain open and refuse close due to the double click function reversing the decision.

Possible fixes:
1. consume the event in the icon clicked to not propagate.
2. stop the triggering in the double click method be checking the target.

i chose n.2 as the request focus is currently handled there and therefore to keep it this way and keep changes to a minimum.