Major changes
Completely rewritten both MFXHLoader and MFXVLoader. Now the LoaderUtils class contains the executor along with the delegate methods to load the FXML file. The views are not loaded automatically by adding them to map but the loading process must be started with the start() method. Only when all the views are loaded the buttons are added to the children list and the default view is loaded. JavaFX tasks are not used anymore because the views are now loaded with a Callable<Node> that returns a Future.
There's also a new option, it's not possible to set the FXMLLoader supplier. For example if you are using a DI framework with JavaFX you probably want to use a FXMLLoader which is managed by the dependency injection.
The support bean MFXLoaderBean has been rewritten to, changed name from MFXLoadItem, knows if the root is the default view to be loaded and provides a nested class Builder which helps to build MFXLoaderBean instances with fluent api.

MFXButton, MFXButtonSkin fixed ripple properties not working properly in SceneBuilder.

Huge refactor for Flowless-based list views.
All list views now extend from a common abstract class, AbstractFlowlessListView which defines all the common properties and common behavior of list views. The class has three parameters: T, C and S. All classes which extend AbstractFlowlessListView are required to specify which cells and selection models are allowed. For example in a check list view I don't want to allow simple cells and a selection model but cells with checkboxes and a check model

About the cells. The first render() is done when the scene is not null. The cells now manage the selection/check behavior and update the model accordingly, this needs to pass the list view instance to the cell on creation. They are not reusable anymore, that means that on scroll new cells are created rather than update old one. The updated index method which gives information about the position of the cell in the virtual flow is now very important and the selection models have been adapted to also keep into account the index. This is necessary for several reasons. For example if you select/check a cell and scroll up/down the cell checks if in the selection model is present its data or index and changes its state accordingly, this allows the selection/check state to survive the scroll. This also helps in case of duplicate data, because one thing for sure is that the index can't be the same.

Because of the changes listed above now the selection/check models have a map property for the data (MapProperty<Integer, T>) backed by a TreeMap

Signed-off-by: PAlex404 <alessandro.parisi406@gmail.com>