New control MFXContextMenu, various fixes and improvements
AbstractFlowlessListView: changed the setItems method because in some situations it's necessary to not clean the selection model (filter combo box for example) so this should be done manually.

AbstractMFXFlowlessListCell: change event handler to filter, this assures that the mouse event is not consumed eventually by other nodes that may be in the cell.

MFXFlowlessListViewSkin: modified the code of VirtualizedScrollPane to make the vBar and hBar protected. MFXVirtualizedScrollPane added methods to get the scroll bars and modified the layoutChildren strategy. The skin now is similar to MFXListViewSkin.

ListSelectionModel: oops, forgot to check if single selection mode.

MFXComboBox, MFXFilterComboBox, MFXComboBoxSkin, MFXFilterComboBoxSkin: switched to MFXFlowlessListView.

ComboSelectionModelMock: clear requested flag is not necessary anymore, update the index too when selecting an item.

MFXButton: fixed ripple generator properties throwing exception because bounded. This is kind of a revert to v11.5.5 but the ripple generator is added in the skin to fix the bug listed in commit 044e6d76.

MFXTableView: fixed behavior when changing items. Allow to change the cells fixed height.
MFXTableViewSkin: big news actually. The layout has been reviewed, when the items list is empty the rows container height is set to a minimum computed by computeRowsContainerHeight(), this also happens when there aren't enough rows to fill the container. The default width is now 600. The pagination controls are now aligned to the left to make them always visible even when the columns are resized. All columns by default have a context menu with the following actions: reset width, reset width of all columns, auto size, auto size all columns.
MFXTableColumnCell: added new custom pseudo class to track drag events.

MFXResourcesLoader: meh, it's more convenient like this.
Signed-off-by: PAlex404 <alessandro.parisi406@gmail.com>