Fixed Dialog Scrim not resizing with pane resize
Made the Dialog Scrim size automatically resizable by binding it to the pane size. Added unbind method to stop scrim resizing in the background after being closed.
Dialog Scrim was previously a static size set on the creation of the dialog, this would not automatically resize it and thus will not cover the entire screen on expanding the window and therefore some components beneath it will be accessible.